feature: prompt for action when on tty
This will prompt for y/N/d (yes, no, details) when running `delete` or
`cleanup` with a non-zero number of results and stdout is a tty.

Yes will delete the thumbnails without having to rerun the search, No
(the default) does nothing, and Details prints the list of files about
to be deleted.

We also change the default for `delete` to be the previous `dry-run`
mode, with `-f/--force` bypassing the prompt (`-d/--dry-run was
removed).